requirements.txt
Flask==0.11.1
flask-mysqldb=0.2.0
Flask-WTF==0.9.4
Flask-Script==0.6.6
Flask-migrate==2.0.0
Flask-Bootstrap==3.0.3.1